requirements.txt
httpx[http2]==0.28.0
pydantic==2.10.0
pydantic-settings==2.6.0
cryptography==43.0.0
python-telegram-bot==21.7
structlog==24.1.0
tenacity==9.0.0
aiosignal==1.3.1
anyio==4.2.0
prometheus-client==0.20.0
pyyaml==6.0.1
aiosqlite==0.20.0
# Optional for Parquet support
# pyarrow==15.0.0
# Testing
pytest==8.4.1
pytest-asyncio==0.24.0
respx==0.21.0
# optional/isolated for Solana live trading (keep commented):
# solders==0.20.0
# solana==0.30.2
